fix: address PR review feedback from Greptile
Extract duplicated resolveCSSVar/resolveColor helpers into shared
color-utils module, add type cast on setViewMode, and add immediate
dispatchAction on merit order chart zoom reset for consistency.
