+++ frontend/src/lib/charts/color-utils.ts
@@ -1,3 +1,21 @@
+/** Resolve a CSS custom property (e.g. `--primary`) to its computed value. */
+export function resolveCSSVar(varName: string): string {
+    if (typeof document === "undefined") return "#888";
+    const root = document.documentElement;
+    let val = getComputedStyle(root).getPropertyValue(varName).trim();
+    const m = val.match(/^var\(([^,)]+)/);
+    if (m?.[1]) val = getComputedStyle(root).getPropertyValue(m[1]).trim();
+    return val || "#888";
+}
+
+/** Resolve a color string that may be a CSS `var(--…)` reference. */
+export function resolveColor(color: string): string {
+    if (!color.startsWith("var(")) return color;
+    const varName = color.match(/^var\(([^,)]+)/)?.[1];
+    if (!varName) return color;
+    return resolveCSSVar(varName);
+}
+
 /** Chart color palette using semantic CSS variables. */
 
 export const CHART_COLORS = [
